py/samples/fastapi-hello/pyproject.toml
[project]
name = "fastapi-hello"
version = "0.1.0"
description = "BugBot: AI Code Reviewer with Genkit + FastAPI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "genkit",
    "genkit-plugin-google-genai",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = ["watchdog>=6.0.0"]

[tool.uv.sources]
genkit = { workspace = true }
genkit-plugin-google-genai = { workspace = true }
